Move gflags from WebRTC into separate repo.
With all the local gflags stuff living in its own repo,
it will be possible to DEPS it in independently of the WebRTC
source code (which also makes it more reusable by other projects).

This change simply adds all the files in
https://chromium.googlesource.com/external/webrtc/+/master/third_party/gflags
(which will be removed as soon this lands and has been properly
added to DEPS).

A codereview.settings file is also added.

BUG=webrtc:5006
TBR=charujain@webrtc.org